Jsmn/ParseError.hpp: Factor out `Jsmn::ParseError` to its own header.
