+++ Jsmn/Parser.cpp
@@ -43,16 +43,6 @@ Jsmn::Detail::Token token_convert(jsmntok_t const& tok) {
 
 namespace Jsmn {
 
-std::string ParseError::enmessage(std::string const& input, unsigned int i) {
-	/* FIXME: Show context around error.  */
-	char cs[2];
-	cs[0] = input[i];
-	cs[1] = 0;
-	return std::string("Parse error near character '") + cs
-	     + std::string("'")
-	     ;
-}
-
 class Parser::Impl {
 private:
 	std::string input;
